finishing cpp09 and all the cpp modules

+++ cpp09/ex01/RPN.hpp
@@ -4,11 +4,12 @@
 #include <string>
 #include <iostream>
 #include <stack>
+#include <stdexcept>
 
 class RPN
 {
     private:
-        std::stack<int> stack; 
+        std::stack<double> stack; 
 
     public:
         RPN();
